fix: move env service to /core
