+++ packages/substream/core/reset.ts
@@ -2,7 +2,7 @@ import { Effect } from 'effect';
 
 import { ipfsCache } from './storage/schema';
 import { Storage, make } from './storage/storage';
-import { Environment, make as makeEnvironment } from '~/sink/environment';
+import { Environment, make as makeEnvironment } from '~/core/environment';
 
 const reset = Effect.gen(function* () {
   const db = yield* Storage;
